requirements.txt
PyQt5~=5.15.11
opencv-python~=4.11.0.86
numpy~=1.26.4
qtawesome~=1.3.1

tensorflow~=2.10.0
matplotlib~=3.7.1
scikit-learn~=1.1.3
pillow~=11.1.0
keyring~=25.6.0
python-dotenv~=1.0.1
cryptography~=44.0.2
argon2-cffi~=23.1.0
